pyproject.toml
[tool.poetry]
name = "xtb_trading_bot"
version = "0.1.0"
description = "xtb trading bot"
authors = ["David Dvořák"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websocket-client = "^1.8.0"
python-dotenv = "^1.0.1"
websockets = "^13.0.1"
matplotlib = "^3.9.2"
mplfinance = "^0.12.10b0"
lightweight-charts = "^2.1"
backtesting = { git = "https://github.com/kernc/backtesting.py.git" }


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
